frontend/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary ###this is for postgree
djangorestframework
django-cors-headers
Bio